context cancelation in normal group tests
